# Gaming.Quiz.API/wwwroot/config/swagger/Gaming.Quiz.API.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Gaming.Quiz.API",
    "description": "The API documentation for Gaming Quiz APIs",
    "version": "V1"
  },
  "paths": {
...
    "/services/Session/login": {
      "post": {
        "tags": [
          "Session"
        ],
        "parameters": [
          {
            "name": "SocialId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "backdoor",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              }
            }
          }
        }
      }
    },
    "/services/Session/anonymouslogin": {
      "post": {
        "tags": [
          "Session"
        ],
        "parameters": [
          {
            "name": "backdoor",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              }
            }
          }
        }
      }
    },
    "/services/Session/user/login": {
      "post": {
        "tags": [
          "Session"
        ],
        "parameters": [
          {
            "name": "waf_guid",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FeedTime": {
        "type": "object",
        "properties": {
          "utcTime": {
            "type": "string",
            "nullable": true
          },
          "istTime": {
            "type": "string",
            "nullable": true
          },
          "cestTime": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HTTPMeta": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "retVal": {
            "type": "integer",
            "format": "int64"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "$ref": "#/components/schemas/FeedTime"
          }
        },
        "additionalProperties": false
      },
      "HTTPResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": false,
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/HTTPMeta"
          }
        },
        "additionalProperties": false
      },
      "Credentials": {
        "type": "object",
        "properties": {
          "optType": {
            "type": "integer",
            "format": "int32"
          },
          "platformId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "socialId": {
            "type": "string",
            "nullable": true
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "emailId": {
            "type": "string",
            "nullable": true
          },
          "countryCode": {
            "type": "string",
            "nullable": true
          },
          "phoneNo": {
            "type": "string",
            "nullable": true
          },
          "profilePicture": {
            "type": "string",
            "nullable": true
          },
          "dob": {
            "type": "string",
            "nullable": true
          },
          "userCreatedDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}